+++ src/lib/acme-core/src/main/java/org/niis/xroad/common/acme/AcmeService.java
@@ -26,65 +26,20 @@
  */
 package org.niis.xroad.common.acme;
 
-import ee.ria.xroad.common.util.AtomicSave;
-
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.jose4j.jws.AlgorithmIdentifiers;
-import org.niis.xroad.common.acme.config.AcmeConfig;
-import org.niis.xroad.common.acme.config.AcmeProperties;
-import org.niis.xroad.common.acme.provider.AcmeCustomSchema;
-import org.niis.xroad.common.exception.BadRequestException;
-import org.niis.xroad.common.vault.AcmeAccountKey;
-import org.niis.xroad.common.vault.VaultClient;
 import org.niis.xroad.globalconf.model.ApprovedCAInfo;
-import org.shredzone.acme4j.Account;
-import org.shredzone.acme4j.AccountBuilder;
-import org.shredzone.acme4j.AcmeJsonResource;
-import org.shredzone.acme4j.Authorization;
-import org.shredzone.acme4j.Certificate;
-import org.shredzone.acme4j.Login;
-import org.shredzone.acme4j.Metadata;
-import org.shredzone.acme4j.Order;
-import org.shredzone.acme4j.RenewalInfo;
-import org.shredzone.acme4j.Session;
-import org.shredzone.acme4j.Status;
-import org.shredzone.acme4j.challenge.Challenge;
-import org.shredzone.acme4j.challenge.Http01Challenge;
-import org.shredzone.acme4j.connector.Resource;
-import org.shredzone.acme4j.exception.AcmeException;
 
-import javax.crypto.spec.SecretKeySpec;
-
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.security.KeyPair;
-import java.security.KeyPairGenerator;
-import java.security.NoSuchAlgorithmException;
-import java.security.SecureRandom;
 import java.security.cert.X509Certificate;
 import java.time.Instant;
-import java.time.Period;
-import java.time.temporal.ChronoUnit;
 import java.util.List;
-import java.util.Optional;
-import java.util.function.Supplier;
 
-@Slf4j
 @RequiredArgsConstructor
 public final class AcmeService {
 
-    public static final int ORDER_NOT_AFTER_DAYS = 365;
-
-    private final AcmeProperties acmeProperties;
-    private final AcmeConfig acmeConfig;
-    private final VaultClient vaultClient;
+    private final AcmeClient acmeClient;
 
     public boolean isExternalAccountBindingRequired(String acmeServerDirectoryUrl) {
-        Session session = new Session(acmeServerDirectoryUrl);
-        return getMetadata(session).isExternalAccountRequired();
+        return acmeClient.isExternalAccountBindingRequired(acmeServerDirectoryUrl);
     }
 
     public List<X509Certificate> orderCertificateFromACMEServer(String commonName,
@@ -94,352 +49,51 @@ public final class AcmeService {
                                                                 String memberId,
                                                                 byte[] certRequest,
                                                                 List<String> contacts) {
-        KeyPair keyPair = getAccountKeyPair(memberId, keyUsage, caInfo);
-        Account account = startSession(keyUsage, caInfo, keyPair, memberId, contacts);
-        Order order = createOrder(commonName, subjectAltName, account);
-        doAuthorizationAndFinalizeOrder(certRequest, order);
+        return acmeClient.orderCertificate(commonName, subjectAltName, toAccountContext(memberId, caInfo, keyUsage, contacts),
+                certRequest);
+    }
 
-        Certificate cert = getCertificate(order);
+    /**
+     * The certificate profile id to send as the ACME {@code profile_id} header for a member-CA order, or
+     * {@code null} if the CA needs none — the {@code keyUsage}-branched signing/authentication profile id.
+     * <p>
+     * Package-private for tests: proves this resolution in isolation, without needing a live ACME server to
+     * observe the header it ultimately drives.
+     */
+    static String resolveCertificateProfileId(ApprovedCAInfo caInfo, AcmeKeyPurpose keyUsage) {
+        return keyUsage == AcmeKeyPurpose.SIGNING
+                ? caInfo.getSigningCertificateProfileId()
+                : caInfo.getAuthenticationCertificateProfileId();
+    }
 
-        return cert != null ? cert.getCertificateChain() : null;
-
+    private AcmeAccountContext toAccountContext(String memberId, ApprovedCAInfo caInfo, AcmeKeyPurpose keyUsage,
+                                                List<String> contacts) {
+        return new AcmeAccountContext(memberId, caInfo.getName(), caInfo.getAcmeServerDirectoryUrl(),
+                resolveCertificateProfileId(caInfo, keyUsage), keyUsage, contacts);
     }
 
     public void checkAccountKeyPairAndRenewIfNecessary(String memberId, ApprovedCAInfo caInfo, AcmeKeyPurpose keyUsage,
                                                         List<String> contacts) {
-        try {
-            String alias = getAlias(memberId, keyUsage, caInfo);
-            Optional<AcmeAccountKey> accountKey = vaultClient.getAcmeAccountKey(alias);
-            if (accountKey.isPresent() && isRenewalDue(accountKey.get())) {
-                Login login = getLogin(memberId, caInfo, keyUsage, contacts);
-                KeyPair keyPair = generateAccountKeyPair();
-                login.getAccount().changeKey(keyPair);
-                vaultClient.createAcmeAccountKey(alias, newAccountKey(keyPair));
-                log.info("Renewed acme account keypair for {}", memberId);
-            }
-        } catch (Exception e) {
-            log.error("Renewing account key pair failed", e);
-        }
-    }
-
-    private boolean isRenewalDue(AcmeAccountKey accountKey) {
-        int renewalTimeBeforeExpirationDate = acmeConfig.getAcmeKeypairRenewalTimeBeforeExpirationDate();
-        return Instant.now().isAfter(accountKey.expiresAt().minus(renewalTimeBeforeExpirationDate, ChronoUnit.DAYS));
-    }
-
-    private KeyPair getAccountKeyPair(String memberId, AcmeKeyPurpose keyUsage, ApprovedCAInfo caInfo) {
-        String alias = getAlias(memberId, keyUsage, caInfo);
-        return vaultClient.getAcmeAccountKey(alias)
-                .map(accountKey -> new KeyPair(accountKey.publicKey(), accountKey.privateKey()))
-                .orElseGet(() -> {
-                    log.debug("Creating keypair for {}", alias);
-                    KeyPair keyPair = generateAccountKeyPair();
-                    vaultClient.createAcmeAccountKey(alias, newAccountKey(keyPair));
-                    return keyPair;
-                });
-    }
-
-    private KeyPair generateAccountKeyPair() {
-        try {
-            KeyPairGenerator keyPairGenerator = KeyPairGenerator.getInstance("RSA");
-            keyPairGenerator.initialize(acmeConfig.getAcmeKeyLength(), new SecureRandom());
-            return keyPairGenerator.generateKeyPair();
-        } catch (NoSuchAlgorithmException e) {
-            throw new AcmeServiceException(e, AcmeDeviationMessage.ACCOUNT_KEY_PAIR_ERROR.build());
-        }
-    }
-
-    private AcmeAccountKey newAccountKey(KeyPair keyPair) {
-        Instant expiresAt = Instant.now().plus(acmeConfig.getAcmeCertificateAccountKeyPairExpiration(), ChronoUnit.DAYS);
-        return new AcmeAccountKey(keyPair.getPrivate(), keyPair.getPublic(), expiresAt);
-    }
-
-    private String getAlias(String memberId, AcmeKeyPurpose keyUsage, ApprovedCAInfo caInfo) {
-        AcmeProperties.Credentials credential = acmeProperties.getEabCredentials(caInfo.getName(), memberId);
-        String alias = memberId;
-        if (credential.getAuthKid() != null && keyUsage == AcmeKeyPurpose.AUTHENTICATION) {
-            alias = "auth_" + alias;
-        } else if (credential.getSignKid() != null && keyUsage == AcmeKeyPurpose.SIGNING) {
-            alias = "sign_" + alias;
-        }
-        return alias;
-    }
-
-    private Account startSession(AcmeKeyPurpose keyUsage, ApprovedCAInfo caInfo, KeyPair keyPair, String memberId,
-                                 List<String> contacts) {
-        try {
-            log.info("Creating session with directory url: {}", caInfo.getAcmeServerDirectoryUrl());
-            String acmeUri;
-            if (caInfo.getAuthenticationCertificateProfileId() != null) {
-                acmeUri = caInfo.getAcmeServerDirectoryUrl().replaceFirst("http", AcmeCustomSchema.XRD_ACME_PROFILE_ID.getSchema());
-            } else {
-                acmeUri = caInfo.getAcmeServerDirectoryUrl().replaceFirst("http", AcmeCustomSchema.XRD_ACME.getSchema());
-            }
-            Session session = new Session(acmeUri);
-            Metadata metadata = getMetadata(session);
-            log.debug("ACME server metadata: {}", metadata.getJSON().toString());
-            log.debug("Creating account");
-            AccountBuilder accountBuilder = new AccountBuilder()
-                    .agreeToTermsOfService()
-                    .useKeyPair(keyPair);
-            addContacts(accountBuilder, contacts);
-            if (metadata.isExternalAccountRequired()) {
-                accountWithEabCredentials(accountBuilder, keyUsage, caInfo, memberId);
-            }
-            return accountBuilder.create(session);
-        } catch (AcmeException e) {
-            throw new AcmeServiceException(e, AcmeDeviationMessage.ACCOUNT_CREATION_FAILURE.build());
-        }
-    }
-
-    private static void addContacts(AccountBuilder accountBuilder, List<String> contacts) {
-        if (contacts != null) {
-            contacts.forEach(accountBuilder::addContact);
-        }
-    }
-
-    private void accountWithEabCredentials(AccountBuilder accountBuilder, AcmeKeyPurpose keyUsage, ApprovedCAInfo caInfo,
-                                           String memberId) {
-        AcmeProperties.Credentials credential = acmeProperties.getEabCredentials(caInfo.getName(), memberId);
-        String kid, secret;
-        if (credential.getAuthKid() != null && keyUsage == AcmeKeyPurpose.AUTHENTICATION) {
-            kid = credential.getAuthKid();
-            secret = credential.getAuthMacKey();
-        } else if (credential.getSignKid() != null && keyUsage == AcmeKeyPurpose.SIGNING) {
-            kid = credential.getSignKid();
-            secret = credential.getSignMacKey();
-        } else {
-            kid = credential.getKid();
-            secret = credential.getMacKey();
-        }
-        if (acmeProperties.isEabMacKeyBase64Encoded(caInfo.getName())) {
-            String secretWithPadding = padBase64(secret);
-            accountBuilder.withKeyIdentifier(kid, secretWithPadding);
-        } else {
-            accountBuilder.withKeyIdentifier(kid, new SecretKeySpec(secret.getBytes(StandardCharsets.UTF_8), "HMAC"));
-        }
-        accountBuilder.withMacAlgorithm(AlgorithmIdentifiers.HMAC_SHA256);
-    }
-
-    private static Metadata getMetadata(Session session) {
-        try {
-            return session.getMetadata();
-        } catch (Exception e) {
-            throw new AcmeServiceException(e, AcmeDeviationMessage.FETCHING_METADATA_ERROR.build());
-        }
-    }
-
-    @SuppressWarnings("checkstyle:MagicNumber")
-    private static String padBase64(String base64StringWithoutPadding) {
-        if (base64StringWithoutPadding.length() % 4 == 3) {
-            return base64StringWithoutPadding + "=";
-        } else if (base64StringWithoutPadding.length() % 4 == 2) {
-            return base64StringWithoutPadding + "==";
-        } else if (base64StringWithoutPadding.length() % 4 == 0) {
-            return base64StringWithoutPadding;
-        } else {
-            throw new BadRequestException(AcmeDeviationMessage.EAB_SECRET_LENGTH.build());
-        }
-    }
-
-    private Order createOrder(String commonName, String subjectAltName, Account account) {
-        try {
-            log.debug("Creating new order");
-            return account.newOrder()
-                    .domains(subjectAltName != null ? subjectAltName : commonName)
-                    .notAfter(Instant.now().plus(Period.ofDays(ORDER_NOT_AFTER_DAYS)))
-                    .create();
-        } catch (AcmeException e) {
-            throw new AcmeServiceException(e, AcmeDeviationMessage.ORDER_CREATION_FAILURE.build());
-        }
-    }
-
-    private void doAuthorizationAndFinalizeOrder(byte[] certRequest, Order order) {
-        log.debug("Starting authorization");
-        for (Authorization auth : order.getAuthorizations()) {
-            if (auth.getStatus() == Status.PENDING) {
-                authorizeAndFinalizeOrder(auth, certRequest, order);
-            }
-        }
-    }
-
-    private void authorizeAndFinalizeOrder(Authorization auth, byte[] certRequest, Order order) {
-        Http01Challenge httpChallenge = auth.findChallenge(Http01Challenge.class)
-                .orElseThrow(() -> new AcmeServiceException(AcmeDeviationMessage.HTTP_CHALLENGE_MISSING.build()));
-        String token = httpChallenge.getToken();
-        if (!AcmeConfig.isValidChallengeToken(token)) {
-            throw new AcmeServiceException(AcmeDeviationMessage.HTTP_CHALLENGE_TOKEN_INVALID.build());
-        }
-        var acmeChallenge = Path.of(acmeConfig.getAcmeChallengePath()).resolve(token);
-        writeChallengeFile(acmeChallenge, httpChallenge.getAuthorization());
-        triggerChallenge(httpChallenge);
-        waitForTheChallengeToBeCompleted(httpChallenge);
-        deleteChallengeFile(acmeChallenge);
-        finalizeOrder(order, certRequest);
-    }
-
-    private void writeChallengeFile(Path acmeChallenge, String content) {
-        try {
-            AtomicSave.execute(acmeChallenge, "tmp_challenge",
-                    out -> out.write(content.getBytes(StandardCharsets.UTF_8)));
-        } catch (Exception e) {
-            throw new AcmeServiceException(e, AcmeDeviationMessage.HTTP_CHALLENGE_FILE_CREATION.build());
-        }
-    }
-
-    private void triggerChallenge(Http01Challenge httpChallenge) {
-        try {
-            httpChallenge.trigger();
-        } catch (AcmeException e) {
-            throw new AcmeServiceException(e, AcmeDeviationMessage.CHALLENGE_TRIGGER_FAILURE.build());
-        }
-    }
-
-    private void deleteChallengeFile(Path acmeChallenge) {
-        try {
-            Files.delete(acmeChallenge);
-        } catch (IOException e) {
-            throw new AcmeServiceException(e, AcmeDeviationMessage.HTTP_CHALLENGE_FILE_DELETION.build());
-        }
-    }
-
-    private void finalizeOrder(Order order, byte[] certRequest) {
-        try {
-            log.debug("Finalizing order");
-            order.execute(certRequest);
-        } catch (AcmeException e) {
-            throw new AcmeServiceException(e, AcmeDeviationMessage.ORDER_FINALIZATION_FAILURE.build());
-        }
-    }
-
-    private void waitForTheChallengeToBeCompleted(Challenge challenge) {
-        log.debug("Waiting for challenge to be completed");
-        int attempts = acmeConfig.getAcmeAuthorizationWaitAttempts();
-        long interval = acmeConfig.getAcmeAuthorizationWaitInterval();
-        waitForTheAcmeResourceToBeCompleted(challenge,
-                challenge::getStatus,
-                attempts,
-                interval,
-                AcmeDeviationMessage.AUTHORIZATION_FAILURE,
-                AcmeDeviationMessage.AUTHORIZATION_WAIT_FAILURE);
-    }
-
-    private static void waitForTheAcmeResourceToBeCompleted(AcmeJsonResource acmeJsonResource,
-                                                            Supplier<Status> statusSupplier,
-                                                            int attempts,
-                                                            long interval,
-                                                            AcmeDeviationMessage fetchFailure,
-                                                            AcmeDeviationMessage fetchWaitFailure) {
-        while (statusSupplier.get() != Status.VALID && attempts-- > 0) {
-            if (statusSupplier.get() == Status.INVALID) {
-                throw new AcmeServiceException(fetchFailure.build());
-            }
-            Instant now = Instant.now();
-            try {
-                Instant retryAfter = acmeJsonResource.fetch().orElse(now.plusSeconds(interval));
-                Thread.sleep(now.until(retryAfter, ChronoUnit.MILLIS));
-            } catch (AcmeException e) {
-                throw new AcmeServiceException(e, fetchFailure.build());
-            } catch (InterruptedException e) {
-                Thread.currentThread().interrupt();
-                throw new AcmeServiceException(e, fetchWaitFailure.build());
-            }
-        }
-        if (statusSupplier.get() != Status.VALID) {
-            throw new AcmeServiceException(fetchWaitFailure.build());
-        }
-    }
-
-    private Certificate getCertificate(Order order) {
-        log.debug("Getting the certificate");
-        int attempts = acmeConfig.getAcmeCertificateWaitAttempts();
-        long interval = acmeConfig.getAcmeCertificateWaitInterval();
-        waitForTheAcmeResourceToBeCompleted(order, order::getStatus, attempts, interval,
-                AcmeDeviationMessage.CERTIFICATE_FAILURE, AcmeDeviationMessage.CERTIFICATE_WAIT_FAILURE);
-        return order.getCertificate();
-    }
-
-    private Login getLogin(String memberId, ApprovedCAInfo approvedCA, AcmeKeyPurpose keyUsage, List<String> contacts) {
-        KeyPair accountKeyPair = getAccountKeyPair(memberId, keyUsage, approvedCA);
-        Session session = new Session(approvedCA.getAcmeServerDirectoryUrl());
-        try {
-            AccountBuilder accountBuilder = new AccountBuilder()
-                    .useKeyPair(accountKeyPair);
-            addContacts(accountBuilder, contacts);
-            accountWithEabCredentials(accountBuilder, keyUsage, approvedCA, memberId);
-            return accountBuilder.createLogin(session);
-        } catch (AcmeException e) {
-            throw new AcmeServiceException(e, AcmeDeviationMessage.ACCOUNT_CREATION_FAILURE.build());
-        }
-    }
-
-    public RenewalInfo getRenewalInfo(String memberId, ApprovedCAInfo approvedCA, X509Certificate certificate,
-                                      AcmeKeyPurpose keyUsage, List<String> contacts) {
-        Login login = getLogin(memberId, approvedCA, keyUsage, contacts);
-        RenewalInfo renewalInfo;
-        try {
-            renewalInfo = login.bindRenewalInfo(certificate);
-            renewalInfo.fetch();
-        } catch (AcmeException e) {
-            throw new AcmeServiceException(e, AcmeDeviationMessage.FETCHING_RENEWAL_INFO_FAILURE.build());
-        }
-        return renewalInfo;
+        acmeClient.checkAccountKeyPairAndRenewIfNecessary(toAccountContext(memberId, caInfo, keyUsage, contacts));
     }
 
     public boolean isRenewalRequired(String memberId, ApprovedCAInfo approvedCA, X509Certificate certificate,
                                      AcmeKeyPurpose keyUsage, List<String> contacts) {
-        return !getRenewalInfo(memberId, approvedCA, certificate, keyUsage, contacts).renewalIsNotRequired(Instant.now());
-    }
-
-    public Instant getSuggestedRenewalStartTime(String memberId,
-                                                ApprovedCAInfo approvedCA,
-                                                X509Certificate certificate,
-                                                AcmeKeyPurpose keyUsage,
-                                                List<String> contacts) {
-        return getRenewalInfo(memberId, approvedCA, certificate, keyUsage, contacts).getSuggestedWindowStart();
+        return acmeClient.isRenewalRequired(toAccountContext(memberId, approvedCA, keyUsage, contacts), certificate);
     }
 
     public Instant getNextRenewalTime(String memberId, ApprovedCAInfo approvedCA, X509Certificate x509Certificate,
                                       AcmeKeyPurpose keyUsage, List<String> contacts) {
-        try {
-            if (hasRenewalInfo(memberId, approvedCA, keyUsage, contacts)) {
-                return getSuggestedRenewalStartTime(memberId, approvedCA, x509Certificate, keyUsage, contacts);
-            }
-        } catch (Exception ex) {
-            log.error(
-                    "Retrieving renewal information from ACME Server failed. "
-                            + "Falling back to fixed renewal time based on certificate expiration date: {}", ex.getMessage());
-        }
-        int renewalTimeBeforeExpirationDate = acmeConfig.getAcmeRenewalTimeBeforeExpirationDate();
-        return x509Certificate.getNotAfter().toInstant().minus(renewalTimeBeforeExpirationDate, ChronoUnit.DAYS);
+        return acmeClient.getNextRenewalTime(toAccountContext(memberId, approvedCA, keyUsage, contacts), x509Certificate);
     }
 
     public boolean hasRenewalInfo(String memberId, ApprovedCAInfo approvedCA, AcmeKeyPurpose keyUsage, List<String> contacts) {
-        try {
-            return getLogin(memberId, approvedCA, keyUsage, contacts).getSession().resourceUrlOptional(Resource.RENEWAL_INFO).isPresent();
-        } catch (AcmeException e) {
-            throw new AcmeServiceException(e, AcmeDeviationMessage.FETCHING_RENEWAL_INFO_FAILURE.build());
-        }
+        return acmeClient.hasRenewalInfo(toAccountContext(memberId, approvedCA, keyUsage, contacts));
     }
 
     public List<X509Certificate> renew(String memberId, String subjectAltName, ApprovedCAInfo approvedCA,
                                        AcmeKeyPurpose keyUsage,
                                        X509Certificate oldCertificate, byte[] newCsr, List<String> contacts) {
-        Login login = getLogin(memberId, approvedCA, keyUsage, contacts);
-        Order order;
-        try {
-            order = login.newOrder()
-                    .domains(subjectAltName)
-                    .notAfter(Instant.now().plus(Period.ofDays(ORDER_NOT_AFTER_DAYS)))
-                    .replaces(oldCertificate)
-                    .create();
-        } catch (AcmeException e) {
-            throw new AcmeServiceException(e, AcmeDeviationMessage.ORDER_CREATION_FAILURE.build());
-        }
-        doAuthorizationAndFinalizeOrder(newCsr, order);
-
-        Certificate cert = getCertificate(order);
-
-        return cert != null ? cert.getCertificateChain() : null;
+        return acmeClient.renew(toAccountContext(memberId, approvedCA, keyUsage, contacts), subjectAltName, oldCertificate, newCsr);
     }
 }
