+++ src/security-server/admin-service/application/src/main/java/org/niis/xroad/securityserver/restapi/dstls/DsTlsAcmeService.java
@@ -26,10 +26,11 @@
 package org.niis.xroad.securityserver.restapi.dstls;
 
 import lombok.RequiredArgsConstructor;
+import org.niis.xroad.common.acme.AcmeAccountContext;
+import org.niis.xroad.common.acme.AcmeClient;
 import org.niis.xroad.common.acme.AcmeKeyPurpose;
 import org.niis.xroad.common.acme.AcmeService;
-import org.niis.xroad.globalconf.model.ApprovedCAInfo;
-import org.niis.xroad.globalconf.model.DsTlsCaInfo;
+import org.niis.xroad.globalconf.model.ApprovedDsTlsCaInfo;
 import org.niis.xroad.securityserver.restapi.config.AdminServiceProperties;
 import org.springframework.stereotype.Component;
 
@@ -40,7 +41,7 @@ import java.util.List;
 import static java.util.Objects.requireNonNullElse;
 
 /**
- * Thin DS TLS-specific wrapper around the shared {@link AcmeService} engine (order, renew, ARI-aware renewal
+ * Thin DS TLS-specific wrapper around the shared {@link AcmeClient} engine (order, renew, ARI-aware renewal
  * timing) — no signer, no member id, no {@code KeyUsageInfo}.
  * <p>
  * Every call is made under a fixed, non-member ACME account alias (see {@link #DS_TLS_ACME_ALIAS}), reusing the
@@ -48,10 +49,10 @@ import static java.util.Objects.requireNonNullElse;
  * The alias can never collide with a real encoded {@code ClientId}, because every encoded {@code ClientId} contains
  * a {@code :} separator and this alias does not.
  * <p>
- * {@link AcmeKeyPurpose#AUTHENTICATION} is passed to the shared engine for every call purely as a required, but
- * here inert, parameter: it only changes which EAB credential sub-fields ({@code auth-kid}/{@code sign-kid}) get
- * selected, and the DS TLS alias's EAB entry carries none of those, so the plain {@code kid}/{@code mac-key} pair
- * is always used regardless of which purpose value is passed.
+ * This class is the only place that knows {@link ApprovedDsTlsCaInfo} exists — it's translated into a plain
+ * {@link AcmeAccountContext} here before ever reaching {@link AcmeClient}, so the shared engine never needs to
+ * know this CA-info shape exists at all. This class depends directly on {@link AcmeClient}, never on
+ * {@link AcmeService} — that class is member auth/sign's own façade over the same shared engine.
  */
 @Component
 @RequiredArgsConstructor
@@ -59,53 +60,40 @@ class DsTlsAcmeService {
 
     static final String DS_TLS_ACME_ALIAS = "dataspace-tls";
 
-    private static final AcmeKeyPurpose DS_TLS_ACME_KEY_PURPOSE = AcmeKeyPurpose.AUTHENTICATION;
     private static final List<String> NO_CONTACTS = List.of();
 
-    private final AcmeService acmeService;
+    private final AcmeClient acmeClient;
     private final AdminServiceProperties adminServiceProperties;
 
     /**
      * Orders a brand-new DS TLS certificate: no certificate exists yet for this CSR's key.
      */
-    List<X509Certificate> enroll(DsTlsCaInfo caInfo, String hostname, byte[] certRequest) {
-        return acmeService.orderCertificateFromACMEServer(hostname, hostname, DS_TLS_ACME_KEY_PURPOSE,
-                toApprovedCaInfo(caInfo), DS_TLS_ACME_ALIAS, certRequest, resolveContacts());
+    List<X509Certificate> enroll(ApprovedDsTlsCaInfo caInfo, String hostname, byte[] certRequest) {
+        return acmeClient.orderCertificate(hostname, hostname, toAccountContext(caInfo), certRequest);
     }
 
     /**
      * Renews the currently stored DS TLS certificate, referencing it via ACME Renewal Information where the CA
      * supports it.
      */
-    List<X509Certificate> renew(DsTlsCaInfo caInfo, String hostname, X509Certificate currentCertificate, byte[] certRequest) {
-        return acmeService.renew(DS_TLS_ACME_ALIAS, hostname, toApprovedCaInfo(caInfo), DS_TLS_ACME_KEY_PURPOSE,
-                currentCertificate, certRequest, resolveContacts());
+    List<X509Certificate> renew(ApprovedDsTlsCaInfo caInfo, String hostname, X509Certificate currentCertificate, byte[] certRequest) {
+        return acmeClient.renew(toAccountContext(caInfo), hostname, currentCertificate, certRequest);
     }
 
     /**
      * The time {@code certificate} is next due for ACME renewal: the CA's suggested ACME Renewal Information window
      * start where available, otherwise a fixed number of days before the certificate's own expiry.
      */
-    Instant getNextRenewalTime(DsTlsCaInfo caInfo, X509Certificate certificate) {
-        return acmeService.getNextRenewalTime(DS_TLS_ACME_ALIAS, toApprovedCaInfo(caInfo), certificate,
-                DS_TLS_ACME_KEY_PURPOSE, resolveContacts());
+    Instant getNextRenewalTime(ApprovedDsTlsCaInfo caInfo, X509Certificate certificate) {
+        return acmeClient.getNextRenewalTime(toAccountContext(caInfo), certificate);
+    }
+
+    private AcmeAccountContext toAccountContext(ApprovedDsTlsCaInfo caInfo) {
+        return new AcmeAccountContext(DS_TLS_ACME_ALIAS, caInfo.getName(), caInfo.getAcmeServerDirectoryUrl(),
+                caInfo.getDsTlsCertificateProfileId(), AcmeKeyPurpose.AUTHENTICATION, resolveContacts());
     }
 
     private List<String> resolveContacts() {
         return requireNonNullElse(adminServiceProperties.getDataspace().getTlsCertificateContacts(), NO_CONTACTS);
-    }
-
-    /**
-     * Adapts a {@link DsTlsCaInfo} globalconf entry to the {@link ApprovedCAInfo} shape the shared ACME engine
-     * expects. {@code authenticationCertificateProfileId} is deliberately left {@code null} regardless of whether
-     * {@code caInfo} carries a {@code dsTlsCertificateProfileId}: that field only ever steers the engine's
-     * profile-id ACME URI scheme, whose header-building step resolves the profile id by matching against
-     * {@code GlobalConfProvider.getApprovedCAs()} (the member CA list) — DS TLS CAs live in a separate list that
-     * mechanism was never wired to consult. Every DS TLS order therefore uses the plain ACME scheme, which covers
-     * every CA this feature targets on day one (public DSP-only CAs need no profile id at all).
-     */
-    private static ApprovedCAInfo toApprovedCaInfo(DsTlsCaInfo caInfo) {
-        return new ApprovedCAInfo(caInfo.getName(), null, null, null,
-                caInfo.getAcmeServerDirectoryUrl(), caInfo.getAcmeServerIpAddress(), null, null);
     }
 }
